Fix a memory leak in the web view
Fix a memory leak in the web view of Nightscout, caused by recreating
SVG elements to display treatments on the timeline, but not removing old
ones. This results in a large number of identical/overlapping SVG
elements, causing a gradual increase in memory usage and decrease in
speed over time, until the tab crashes (which takes about a day for me).